requirements.txt
Flask>=2.0.0
requests>=2.28.0
python-dotenv>=1.0.0
Flask-Limiter==3.5.0
PyYAML>=6.0
